+++ core/components/big-data-list/big-data-list.vue
@@ -144,14 +144,10 @@ export default {
 	},
 	methods: {
 		defaultSelected(callback) {
-			if (
-				this.value === null ||
-				this.value === undefined ||
-				!this.value.toString()
-			) {
-				this.loading = false
-				return
-			}
+			this.currentValueArr = []
+			this.query = ''
+			this.currentItemArr = []
+			this.loading = false
 			this.value instanceof Array
 				? (this.currentValueArr = this.value)
 				: (this.currentValueArr[0] = this.value)