+++ core/components/time-picker/time-picker.vue
@@ -18,44 +18,40 @@
 					:clearable="clearable"
 					v-model="currentValue"
 					:dropShow="dropShow"
-					@on-remove-tag="
-						'select'
-
-
-					"
 					@on-clear="clearTime"
 					@click.native="!disabled && dropShowFocus()"
 				/>
 			</template>
 			<TimeSpinner
 				:dropShow="dropShow"
-				:format="format"
 				:steps="steps"
 				:clock-column="clockColumn"
 				:disabled-times="disabledTimes"
 				:hide-disabled-options="hideDisabledOptions"
 				:use12-hour-system="use12HourSystem"
 				v-model="time"
+				@on-select="select"
 			/>
 		</Drop>
 	</div>
 </template>
 <script>
 import Drop from '../drop'
+import TimeSpinner from './time-spinner'
 import SelectHead from '../select/select-head'
 import { UseSelectHeaderMixin } from '../../mixins'
 import TimePickerMixin from './time-picker-mixin'
 import dayjs from 'dayjs'
 import toArray from 'dayjs/plugin/toArray'
 dayjs.extend(toArray)
-let now = dayjs()
 const prefix = 'zov-time-picker'
 export default {
 	name: prefix,
 	mixins: [UseSelectHeaderMixin, TimePickerMixin],
 	components: {
 		Drop,
-		SelectHead
+		SelectHead,
+		TimeSpinner
 	},
 	props: {
 		// 组件组装
@@ -95,50 +91,64 @@ export default {
 			default() {
 				return this.use12HourSystem ? 'hh:mm:ss a' : 'HH:mm:ss'
 			}
+		},
+		valueFormat: {
+			type: String,
+			default() {
+				return this.use12HourSystem ? 'hh:mm:ss a' : 'HH:mm:ss'
+			}
 		}
 	},
 	data() {
 		return {
 			currentValue: '',
-			dropShow: false
+			dropShow: false,
+			actionExecution: false
 		}
 	},
 	watch: {
 		dropShow(val) {
 			this.$emit('on-open-change', val)
 		},
+		value(val) {
+			if (this.actionExecution) {
+				return
+			}
+			this.time = this.formatToArr(this.parse(val))
+			this.select()
+		},
 		time(val) {
-			this.currentValue = val.length ? this.arrToFormat(val) : ''
-			this.$emit('input', this.currentValue)
-			this.$emit('on-change', this.currentValue)
+			this.currentValue = this.arrToFormat(val)
 		}
 	},
 	methods: {
+		select() {
+			this.actionExecution = true
+			this.$emit('input', this.arrToFormat(this.time, this.valueFormat))
+			this.$emit(
+				'on-change',
+				this.arrToFormat(this.time, this.valueFormat)
+			)
+			this.$nextTick(() => {
+				this.actionExecution = false
+			})
+		},
 		parse(val) {
-			return dayjs('1990-05-04 ' + val)
+			return dayjs('2000-00-00 ' + val)
 		},
 		formatToArr(date) {
 			return date.toArray().slice(3, 6)
 		},
-		arrToFormat(arr) {
-			return this.parse(arr.join(':')).format(this.format)
+		arrToFormat(arr, format) {
+			format = format || this.format
+			return arr.length ? this.parse(arr.join(':')).format(format) : ''
 		},
 		clearTime() {
 			this.time = []
 		}
 	},
 	mounted() {
-		if (
-			!(
-				this.value === null ||
-				this.value === undefined ||
-				!this.value.toString()
-			)
-		) {
-			this.time = this.value
-				? this.formatToArr(this.parse(this.value))
-				: this.formatToArr(now)
-		}
+		this.time = this.value ? this.formatToArr(this.parse(this.value)) : []
 		this.autofocus ? this.dropShowFocus() : this.dropHideBlur()
 	}
 }