+++ CHANGELOG.md
@@ -2,8 +2,23 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.11.3"></a>
+## [0.11.3](https://github.com/zovui/zov/compare/v0.11.1...v0.11.3) (2020-04-14)
+
+
+### Bug Fixes
+
+* **select:** value change ([9f3f491](https://github.com/zovui/zov/commit/9f3f491))
+
+
+
 <a name="0.11.2"></a>
-## [0.11.2](https://github.com/zovui/zov/compare/v0.11.1...v0.11.2) (2019-11-22)
+## [0.11.2](https://github.com/zovui/zov/compare/v0.11.1...v0.11.2) (2020-04-14)
+
+
+### Bug Fixes
+
+* **select:** value change ([9f3f491](https://github.com/zovui/zov/commit/9f3f491))